requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
idna==2.9
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
pycparser==2.19
pyzmq==19.0.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
torchnet==0.0.4
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
urllib3==1.25.8
visdom==0.1.8.9
websocket-client==0.57.0
wincertstore==0.2
